requirements.txt
requests>=2.31.0
httpx>=0.25.0
playwright>=1.40.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.2.0
